# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: VPS Billing Platform API
  version: 2.0.0
  license: {name: Proprietary}
servers:
  - url: /api/v1
security: [{userSession: []}]
paths:
...
  /orders:
    get:
      summary: List user orders
      operationId: listOrders
      responses:
        "200": {description: Orders}
        "401": {$ref: "#/components/responses/Unauthenticated"}
    post:
      summary: Create an order
      operationId: createOrder
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan_id, quantity]
              properties:
                plan_id: {type: string, format: uuid}
                quantity: {type: integer, minimum: 1}
                promotion_code: {type: string, maxLength: 64}
      responses:
        "201": {description: "Order, invoice, and pending payment created"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "422": {$ref: "#/components/responses/ValidationFailure"}
...
  /subscriptions:
    get:
      summary: List user subscriptions
      operationId: listSubscriptions
      responses:
        "200":
          description: Subscriptions
          content:
            application/json:
              schema:
                type: object
                required: [success, data, request_id]
                properties:
                  success: {type: boolean, const: true}
                  request_id: {type: string}
                  data:
                    type: object
                    required: [items]
                    properties:
                      items: {type: array, items: {$ref: "#/components/schemas/Subscription"}}
        "401": {$ref: "#/components/responses/Unauthenticated"}
...
  /instances/{id}:
    get:
      summary: Get an instance
      operationId: getInstance
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": {description: Instance}
        "404": {description: Not found}
...
  /tickets:
    get:
      summary: List current user support tickets
      operationId: listTickets
      responses:
        "200": {description: Tickets}
    post:
      summary: Create a support ticket with its first message
      operationId: createTicket
      parameters: [{$ref: "#/components/parameters/CSRFToken"}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject, priority, message]
              additionalProperties: false
              properties:
                subject: {type: string, minLength: 3, maxLength: 255}
                priority: {type: string, enum: [low, normal, high]}
                message: {type: string, minLength: 1, maxLength: 10000}
      responses:
        "201": {description: Ticket created}
        "422": {$ref: "#/components/responses/ValidationFailure"}
...
  /tickets/{id}/messages:
    post:
      summary: Add a user message to an owned support ticket
      operationId: addTicketMessage
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              additionalProperties: false
              properties: {message: {type: string, minLength: 1, maxLength: 10000}}
      responses:
        "201": {description: Message created}
        "404": {$ref: "#/components/responses/NotFound"}
  /operations/{id}:
    get:
      summary: Get an operation
      operationId: getOperation
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Operation
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OperationResponse"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "404": {$ref: "#/components/responses/NotFound"}
...
  /admin/dashboard:
    get:
      summary: Get anomaly-first administrator dashboard
      security: [{adminSession: []}]
      responses: {"200": {description: Health, alerts, capacity, providers, and nodes}, "403": {description: Permission denied}}
  /admin/users:
    get: {summary: List users, security: [{adminSession: []}], responses: {"200": {description: Users}, "403": {description: Permission denied}}}
  /admin/users/{id}/status:
    patch: {summary: Suspend or activate a user with audit, security: [{adminSession: []}], parameters: [{$ref: "#/components/parameters/ResourceID"}], responses: {"200": {description: Updated user}, "403": {description: Permission denied}}}
  /admin/products:
    get: {summary: List products and plans, security: [{adminSession: []}], responses: {"200": {description: Products}}}
  /admin/orders:
    get: {summary: List orders, security: [{adminSession: []}], responses: {"200": {description: Orders}}}
  /admin/payments:
    get: {summary: List payments, security: [{adminSession: []}], responses: {"200": {description: Payments}}}
  /admin/ledger:
    get: {summary: List immutable ledger transactions, security: [{adminSession: []}], responses: {"200": {description: Ledger transactions}}}
  /admin/ledger/adjustments:
    post: {summary: Create an audited balanced wallet adjustment, security: [{adminSession: []}], responses: {"201": {description: Wallet projection and transaction ID}, "409": {description: Insufficient balance}}}
  /admin/subscriptions:
    get: {summary: List subscriptions, security: [{adminSession: []}], responses: {"200": {description: Subscriptions}}}
  /admin/instances:
    get: {summary: List instances, security: [{adminSession: []}], responses: {"200": {description: Instances}}}
  /admin/instances/{id}:
    get: {summary: Get instance operations and diagnostics subject to RBAC, security: [{adminSession: []}], parameters: [{$ref: "#/components/parameters/ResourceID"}], responses: {"200": {description: Instance detail}, "404": {description: Not found}}}
  /admin/nodes:
    get: {summary: List nodes and capacity, security: [{adminSession: []}], responses: {"200": {description: Nodes}}}
  /admin/providers:
    get: {summary: List providers without credentials, security: [{adminSession: []}], responses: {"200": {description: Providers}}}
  /admin/operations:
    get: {summary: List operations, security: [{adminSession: []}], responses: {"200": {description: Operations}}}
  /admin/operations/{id}:
    get: {summary: Get workflow steps and RBAC-filtered raw diagnostics, security: [{adminSession: []}], parameters: [{$ref: "#/components/parameters/ResourceID"}], responses: {"200": {description: Operation detail}, "404": {description: Not found}}}
  /admin/tickets:
    get: {summary: List support tickets, security: [{adminSession: []}], responses: {"200": {description: Tickets}}}
  /admin/tickets/{id}/messages:
    post: {summary: Add an audited administrator reply, security: [{adminSession: []}], parameters: [{$ref: "#/components/parameters/ResourceID"}], responses: {"201": {description: Ticket message}, "409": {description: Ticket finalized}}}
  /admin/tickets/{id}/status:
    patch: {summary: Change ticket state with audit, security: [{adminSession: []}], parameters: [{$ref: "#/components/parameters/ResourceID"}], responses: {"200": {description: Updated ticket}}}
  /admin/audit:
    get: {summary: List audit events, security: [{adminSession: []}], responses: {"200": {description: Audit events}}}
  /admin/admins:
    get: {summary: List administrators and roles, security: [{adminSession: []}], responses: {"200": {description: Administrators}}}
  /admin/admins/{id}/roles:
    put: {summary: Replace administrator roles with audit, security: [{adminSession: []}], parameters: [{$ref: "#/components/parameters/ResourceID"}], responses: {"200": {description: Updated administrator roles}}}
  /admin/roles:
    get: {summary: List roles and permissions, security: [{adminSession: []}], responses: {"200": {description: Roles}}}
  /admin/settings:
    get: {summary: List settings with secret values masked, security: [{adminSession: []}], responses: {"200": {description: Settings}}}
  /admin/settings/{key}:
    put:
      summary: Upsert an audited setting
      security: [{adminSession: []}]
      parameters: [{name: key, in: path, required: true, schema: {type: string}}]
      responses: {"200": {description: Masked setting}}
...
  /instances/{id}/usage:
    get: {summary: Get current usage allowance and estimated charge, parameters: [{$ref: "#/components/parameters/Id"}], responses: {"200": {description: Usage summary}, "404": {$ref: "#/components/responses/NotFound"}}}
...
  /admin/providers/{id}:
    get: {summary: Get secret-safe provider health and inventory detail, security: [{adminSession: []}], parameters: [{$ref: "#/components/parameters/ResourceID"}], responses: {"200": {description: Provider detail}, "403": {$ref: "#/components/responses/Forbidden"}}}
  /admin/operations/{id}/retry:
    post: {summary: Create an audited child retry operation, security: [{adminSession: []}], parameters: [{$ref: "#/components/parameters/ResourceID"}, {$ref: "#/components/parameters/CSRFToken"}], responses: {"202": {description: Retry accepted}, "409": {$ref: "#/components/responses/Conflict"}}}
  /admin/operations/{id}/cancel:
    post: {summary: Cancel an operation that has not started, security: [{adminSession: []}], parameters: [{$ref: "#/components/parameters/ResourceID"}, {$ref: "#/components/parameters/CSRFToken"}], responses: {"200": {description: Cancelled operation}, "409": {$ref: "#/components/responses/Conflict"}}}
  /admin/usage:
    get: {summary: List rated usage periods and revenue diagnostics, security: [{adminSession: []}], responses: {"200": {description: Usage periods}, "403": {$ref: "#/components/responses/Forbidden"}}}
  /admin/dead-letters:
    get: {summary: List secret-safe dead-letter events, security: [{adminSession: []}], responses: {"200": {description: Dead letters}}}
  /admin/dead-letters/{id}/replay:
    post: {summary: Replay a dead-letter event with audit, security: [{adminSession: []}], parameters: [{$ref: "#/components/parameters/ResourceID"}, {$ref: "#/components/parameters/CSRFToken"}], responses: {"202": {description: Replay accepted}, "403": {$ref: "#/components/responses/Forbidden"}}}
  /admin/payments/{id}/refunds:
    post: {summary: Create an idempotent ledger-backed refund, security: [{adminSession: []}], parameters: [{$ref: "#/components/parameters/ResourceID"}, {$ref: "#/components/parameters/CSRFToken"}], responses: {"201": {description: Refund created}, "409": {$ref: "#/components/responses/Conflict"}}}
components:
  securitySchemes:
    userSession: {type: apiKey, in: cookie, name: vps_user_session}
    adminSession: {type: apiKey, in: cookie, name: vps_admin_session}
...
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
    Id:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 8, maxLength: 255}
...
  responses:
    Unauthenticated: {description: Authentication required, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
    Forbidden: {description: Permission or CSRF validation failed, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
    Conflict: {description: Resource conflict, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
    ValidationFailure: {description: Validation failure, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
...
    NotFound: {description: Resource not found, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error, request_id]
      properties:
        success: {type: boolean, const: false}
        request_id: {type: string}
        error:
          type: object
          required: [code, message_key]
          properties:
            code: {type: string}
            message_key: {type: string}
            details: {type: object, additionalProperties: true}
...
    Subscription:
      type: object
      required: [id, plan_id, status, billing_cycle, price_minor, currency, cancel_at_period_end, version]
      properties:
        id: {type: string, format: uuid}
        plan_id: {type: string, format: uuid}
        plan_slug: {type: string}
        plan_name_i18n: {type: object, additionalProperties: {type: string}}
        status: {type: string, enum: [pending, active, past_due, suspended, cancelled, expired, terminated]}
        billing_cycle: {type: string, enum: [monthly, quarterly, yearly]}
        price_minor: {type: integer, format: int64}
        currency: {type: string, minLength: 3, maxLength: 3}
        started_at: {type: [string, "null"], format: date-time}
        current_period_start: {type: [string, "null"], format: date-time}
        current_period_end: {type: [string, "null"], format: date-time}
        next_due_at: {type: [string, "null"], format: date-time}
        grace_until: {type: [string, "null"], format: date-time}
        cancel_at_period_end: {type: boolean}
        ended_at: {type: [string, "null"], format: date-time}
        version: {type: integer, format: int64}
...
    OperationResponse:
      type: object
      required: [success, data, request_id]
      properties:
        success: {type: boolean, const: true}
        request_id: {type: string}
        data:
          type: object
          required: [id, type, resource_type, resource_id, status, progress, retryable, retry_count, max_retries, trace_id, created_at, updated_at, steps]
          properties:
            id: {type: string, format: uuid}
            type: {type: string}
            resource_type: {type: string}
            resource_id: {type: string, format: uuid}
            status:
              type: string
              enum: [queued, running, waiting_provider, waiting_resource, verifying, retrying, succeeded, failed, cancelled]
            phase: {type: [string, "null"]}
            progress: {type: integer, minimum: 0, maximum: 100}
            message_key: {type: [string, "null"]}
            retryable: {type: boolean}
            retry_count: {type: integer, minimum: 0}
            max_retries: {type: integer, minimum: 0}
            error_code: {type: [string, "null"]}
            trace_id: {type: string}
            started_at: {type: [string, "null"], format: date-time}
            finished_at: {type: [string, "null"], format: date-time}
            created_at: {type: string, format: date-time}
            updated_at: {type: string, format: date-time}
            steps:
              type: array
              items: {$ref: "#/components/schemas/OperationStep"}
    OperationStep:
      type: object
      required: [key, order, status, progress, attempt]
      properties:
        key: {type: string}
        order: {type: integer}
        status: {type: string, enum: [pending, running, waiting, succeeded, failed, skipped]}
        progress: {type: integer, minimum: 0, maximum: 100}
        attempt: {type: integer, minimum: 0}
        error_code: {type: [string, "null"]}
        output: {type: object, additionalProperties: true}
        started_at: {type: [string, "null"], format: date-time}
        finished_at: {type: [string, "null"], format: date-time}